refactor(code): split the code to improve organization
Split the code into
- appinio_swiper
- appinio_swiper_controller
- appinio_unswipe_card
- enums
- typedefs
