+++ packages/appinio_swiper/lib/src/enums.dart
@@ -1,0 +1,5 @@
+enum AppinioSwiperState { swipe, swipeLeft, swipeRight, unswipe }
+
+enum AppinioSwiperDirection { none, left, right, top, bottom }
+
+enum AppinioSwiperSwipeType { none, swipe, unswipe, goBack }
